Only install the AlertManager plugin and datasource in Grafana when AlertManager has been enabled in the Prometheus service instance (#4)
* Implemented logic to check if AlertManager in the Prometheus service instance has been enabled first before installing the AlertManager plugin and datasource in Grafana


There's a bug in deploying grafana in the current logic, if AlertManager for Prometheus in the service instance has not been enabled. 

Error is AlertManager datasource yaml "yaml: unmarshal errors:\n line 10: cannot unmarshal !!str null into values.StringValue". 

The current logic generates a datasource yaml for AlertManager that has a null value for the datasource URL, when AlertManager in the Prometheus service instance has not been enabled. Then Grafana tries to install AlertManager with this datasource yaml, causing the resulting error. 

The error can be mitigated by checking if AlertManager in the Prometheus service instance exists first before executing the rest of the logic. Else, the AlertManager datasource yaml is not generated.